requirements.txt
C:/luneth_engine
C:/winmode
pygame
pymunk
gif_pygame
numpy
setuptools
screeninfo
Pyinstaller
pygame
pillow





